fix: resolve compiler warnings in engine code
- Fix signedness comparison warnings (-Wsign-compare)
  - globals.cc: use size_t for loop iterating NB_POLICY_VALUES()
  - benchmark.cc: use size_t for loop iterating NB_INPUT_VALUES()
  - training_data_writer.cc: use size_t for loop and channel variable
- Fix deprecated enum-float arithmetic warnings
  - evaluate.cpp: cast PawnValueEg to int before division
  - uci.cpp: cast PawnValueEg to int before division
- Fix unused variable warning in test_move_gen.cc
  - Remove unused moves_a_initial variable

+++ engine/src/globals.cc
@@ -18,7 +18,7 @@ LogLevel parseLogLevel(const std::string& str) {
 }
 
 void init_policy_index() {
-    for (int i = 0; i < NB_POLICY_VALUES(); i++) {
+    for (size_t i = 0; i < NB_POLICY_VALUES(); i++) {
         if (POLICY_INDEX.find(UCI_MOVES[i]) == POLICY_INDEX.end()) {
             POLICY_INDEX[UCI_MOVES[i]] = i; 
         }
